Consolidate in one function
